Export programmatic interface
