+++ src/renderers/swift/_types.ts
@@ -1,4 +1,4 @@
-import type { SwiftGenerationTarget } from '../../api.js';
+import type { SwiftGenerationTarget } from '../../api/index.js';
 import type { SwiftGeneration } from '../../generators/swift/index.js';
 import type { RenderedFile } from '../_types.js';
